+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Fix issue with showing range panel in Images Explorer (roubkar)
+
 ## 3.3.2 Dec 20 2021
 
 - Fix issue with not providing point density value to live-update query (rubenaprikyan)